app/requirements.txt
flask
psycopg2-binary
prometheus_flask_exporter
prometheus_client
